# back/src/app/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Insumos API",
    "version": "1.0.0",
    "description": "Documentación Insumos API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "required": [
          "usuario",
          "password",
          "email",
          "rol"
        ],
        "properties": {
          "usuario_id": {
            "type": "integer",
            "description": "ID autogenerado para el usuario"
          },
          "usuario": {
            "type": "string",
            "description": "Nombre de usuario"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario"
          },
          "email": {
            "type": "string",
            "description": "Dirección de correo electrónico del usuario"
          },
          "rol": {
            "type": "integer",
            "description": "Rol del usuario en el sistema"
          }
        },
        "example": {
          "usuario_id": 1,
          "usuario": "usuarioejemplo",
          "password": "contraseña123",
          "email": "usuario@ejemplo.com",
          "rol": 2
        }
      }
    }
  },
  "paths": {
...
    "/usuarios/login": {
      "post": {
        "summary": "Iniciar sesión",
        "description": "Permite a un usuario iniciar sesión en el sistema.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usuario o email",
                  "password"
                ],
                "properties": {
                  "usuario": {
                    "type": "string",
                    "description": "El nombre de usuario para iniciar sesión"
                  },
                  "email": {
                    "type": "string",
                    "description": "El email del usuario para iniciar sesión"
                  },
                  "password": {
                    "type": "string",
                    "description": "La contraseña del usuario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Token de acceso generado para el usuario"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Autenticación fallida"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/usuarios/crearUsuario": {
      "post": {
        "summary": "Crea un nuevo usuario",
        "description": "Añade un nuevo usuario al sistema.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usuario",
                  "password",
                  "email",
                  "rol"
                ],
                "properties": {
                  "usuario": {
                    "type": "string",
                    "description": "Nombre de usuario para el nuevo usuario"
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña para el nuevo usuario"
                  },
                  "email": {
                    "type": "string",
                    "description": "Dirección de correo electrónico del usuario"
                  },
                  "rol": {
                    "type": "string",
                    "description": "Rol del usuario en el sistema"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/usuarios/obtenerUsuarioPorID/{usuario_id}": {
      "get": {
        "summary": "Obtiene un usuario por su ID",
        "description": "Devuelve los detalles de un usuario específico basado en su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID único del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/usuarios/obtenerTodosLosUsuarios": {
      "get": {
        "summary": "Obtiene todos los usuarios",
        "description": "Devuelve un listado de todos los usuarios registrados.",
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  },
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/usuarios/actualizarUsuario": {
      "put": {
        "summary": "Actualiza un usuario",
        "description": "Actualiza los datos de un usuario existente.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usuario_id"
                ],
                "properties": {
                  "usuario_id": {
                    "type": "integer",
                    "description": "ID del usuario a actualizar"
                  },
                  "usuario": {
                    "type": "string",
                    "description": "Nombre de usuario actualizado"
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña actualizada"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email actualizado"
                  },
                  "rol": {
                    "type": "string",
                    "description": "Rol actualizado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/usuarios/actualizarUsuarioPorID/{usuario_id}": {
      "put": {
        "summary": "Actualiza un usuario por su ID",
        "description": "Actualiza los detalles de un usuario específico basado en su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID único del usuario a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string",
                    "description": "Nombre de usuario actualizado"
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña actualizada"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email actualizado"
                  },
                  "rol": {
                    "type": "string",
                    "description": "Rol actualizado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente"
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/usuarios/eliminarUsuario": {
      "delete": {
        "summary": "Elimina un usuario",
        "description": "Elimina un usuario de la base de datos.",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del usuario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "204": {
            "description": "Usuario eliminado correctamente"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/usuarios/actualizarPass": {
      "put": {
        "summary": "Actualiza la contraseña de un usuario",
        "description": "Permite a un usuario actualizar su contraseña.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usuario_id",
                  "passwordActual",
                  "passwordNuevo"
                ],
                "properties": {
                  "usuario_id": {
                    "type": "integer",
                    "description": "ID del usuario que actualiza su contraseña"
                  },
                  "passwordActual": {
                    "type": "string",
                    "description": "Contraseña actual del usuario"
                  },
                  "passwordNuevo": {
                    "type": "string",
                    "description": "Nueva contraseña del usuario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada exitosamente"
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
...
  },
...
}